requirement.txt
accelerate==0.23.0
matplotlib==3.8.0
nibabel==5.1.0
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.1
pillow==10.0.1
scipy==1.11.3
scikit-image==0.22.0
scikit-learn==1.3.1
pytorch==1.12.1
torchvision==0.13.1
tqdm==4.66.1